Support matching arbitrary patterns in arbitrary (logfile, posfile) pairs
This allows to scan for several independent things in a set of logfiles at
once, with multiple position tracking files for a single logfile (usually
scanning for independent, disjoint events that show up as mixed log entries).
For example, with this log file:

  proc1: started
  proc1: stopped
  proc2: started
  proc1: started
  proc2: stopped
  proc1: stopped
  proc2: started
  proc2: stopped

The following command can be used to scan for the next pair of "started"
entries no matter how the entries for proc1 and proc2 are mixed; see tests/all
for an example.

  logscan --verbose -y started \
      log --label proc1 -p pos1 -f proc1 \
      log --label proc2 -p pos2 -f proc2