Fix sign extension bug on 32-bit platforms
In seek_logfiles(), we were trying to compute the minimum offset to seek to for
all position tracking files (or the beginning of the file without a position
tracking file).  The initial maximum value was computed wrong on platforms
where sizeof(off_t) != sizeof(long).  Fix this by avoiding the need for an
initial maximum value.

Also, we don't actually need to special case expressions that are not covered
by a position tracking file because expr->line and expr->offset are already
initialized to the beginning of the file, so remove that code.